requirements.txt
scikit-learn==1.0.2
pandas==1.1.5
hydra-core==1.1.2
omegaconf==2.1.2
pytest==7.1.2
marshmallow-dataclass==8.5.8
Faker==13.7.0
flake8==4.0.1
